search component moved
